EG-2754, EG-2755 Addressed PR comments
