+++ business-networks/workflows/src/test/kotlin/net/corda/bn/flows/CreateBusinessNetworkFlowTest.kt
@@ -17,7 +17,8 @@ class CreateBusinessNetworkFlowTest : MembershipManagementFlowTest(numberOfAutho
         val networkId = UniqueIdentifier()
 
         runCreateBusinessNetworkFlow(authorisedMember, networkId = networkId)
-        assertFailsWith<DuplicateBusinessNetworkException> { runCreateBusinessNetworkFlow(authorisedMember, networkId = networkId) }
+        val e = assertFailsWith<DuplicateBusinessNetworkException> { runCreateBusinessNetworkFlow(authorisedMember, networkId = networkId) }
+        assertEquals(networkId, e.networkId)
     }
 
     @Test(timeout = 300_000)
@@ -34,7 +35,8 @@ class CreateBusinessNetworkFlowTest : MembershipManagementFlowTest(numberOfAutho
         val networkId = (runCreateBusinessNetworkFlow(authorisedMember).tx.outputStates.single() as MembershipState).networkId
         val groupId = getAllGroupsFromVault(authorisedMember, networkId).single().linearId
 
-        assertFailsWith<DuplicateBusinessNetworkGroupException> { runCreateBusinessNetworkFlow(authorisedMember, groupId = groupId) }
+        val e = assertFailsWith<DuplicateBusinessNetworkGroupException> { runCreateBusinessNetworkFlow(authorisedMember, groupId = groupId) }
+        assertEquals(groupId, e.groupId)
     }
 
     @Test(timeout = 300_000)
